redesign setting menu using mui components
