+++ src/components/ConnectWallet.module.css
@@ -1,5 +1,6 @@
 .wrapper {
   display: flex;
+  flex-direction: column;
   align-items: center;
   justify-content: center;
   /* margin-bottom: 1rem; */
