--- styles/LanguageSelector.module.css
@@ -1,3 +1,0 @@
-.wrapper {
-  margin-bottom: 2rem;
-}
